FreelanceOS/freelanceos/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pydantic>=2.11.7
pydantic-settings>=2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
reportlab==4.1.0
python-dotenv==1.0.1
aiofiles==23.2.1
resend==0.8.0
psycopg2-binary==2.9.9
alembic==1.13.1
supabase==2.28.3
apscheduler==3.11.2
websockets==14.0
